Fix treeview dir toggle producing unexpected state
Currently when toggling the selection state of a directory that contains
partially selected files or lines the recursive contents of the
directory will simply be inverted.

This is rather unexpected as generally one would want the directory to
be set to a consistent state which is dependent on its starting state.

This change corrects this behaviour issue by ensuring the directory
contents is either all selected or all deselected after toggling.