Interleave hunk lines when applying diffs
This fixes a bug where line ordering in a partially selected left/right
hunk was not being maintained.

This was an issue because the old diff application logic would apply all
of the left side of a hunk and then all of the right.

This reworks this logic to apply hunks line-by-line, left to right. This
means that given a hunk spanning multiple lines, for each line the left
will be applied and then the right.

This interleaving results in the expected line ordering being maintained
when partially selecting lines in the same hunk.